refactor: post기능에 SecurityContext 통해 사용자 정보 주입

+++ src/main/java/com/team1/mixIt/post/service/PostService.java
@@ -7,6 +7,7 @@ import com.team1.mixIt.post.dto.response.PostResponse;
 import com.team1.mixIt.post.entity.Post;
 import com.team1.mixIt.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,8 +24,6 @@ public class PostService {
 
     @Transactional
     public void createPost(Long userId, PostCreateRequest request) {
-        // 테스트용 현재 로그인한 사용자 ID는 1로 하드코딩
-        // TODO: 실제 로그인 유저 ID 로 교체
         Post post = Post.builder()
                 .userId(userId)
                 .category(request.getCategory())
@@ -38,7 +37,7 @@ public class PostService {
     @Transactional(readOnly = true)
     public PostResponse getPostById(Long id) {
         Post post = postRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("The post does not exist."));
+                .orElseThrow(() -> new RuntimeException("게시물이 없습니다."));
         return mapToPostResponse(post);
     }
 
@@ -48,22 +47,30 @@ public class PostService {
         return posts.stream().map(this::mapToPostResponse).collect(Collectors.toList());
     }
 
-    @Transactional
-    public void updatePost(Long userId, Long postId, PostUpdateRequest request) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new RuntimeException("The post does not exist."));
-        post.setTitle(request.getTitle());
-        post.setContent(request.getContent());
-        post.setImageIds(nonNull(request.getImageIds()) ? request.getImageIds() : List.of());
 
+    @Transactional
+    public void updatePost(Long userId, Long postId, PostUpdateRequest req) {
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new RuntimeException("게시물이 없습니다."));
+        if (!post.getUserId().equals(userId)) {
+            throw new AccessDeniedException("내 글만 수정할 수 있습니다.");
+        }
+        post.setTitle(req.getTitle());
+        post.setContent(req.getContent());
+        post.setImageIds(nonNull(req.getImageIds()) ? req.getImageIds() : List.of());
     }
+
 
     @Transactional
     public void deletePost(Long userId, Long postId) {
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new RuntimeException("The post does not exist."));
+                .orElseThrow(() -> new RuntimeException("게시물이 없습니다."));
+        if (!post.getUserId().equals(userId)) {
+            throw new AccessDeniedException("내 글만 삭제할 수 있습니다.");
+        }
         postRepository.delete(post);
     }
+
 
     private PostResponse mapToPostResponse(Post post) {
         PostResponse response = new PostResponse();
